python/bin/question_worker/pyproject.toml
[tool.poetry]
name = "question_worker"
version = "0.1.0"
description = ""
authors = ["Nikolay Vasilev <vasilev.nk@phystech.edu>"]

[tool.poetry.dependencies]
python = "^3.9"
nltk = "^3.6.7"
gensim = "^4.1.2"
pymorphy2 = "^0.9.1"
aioredis = "^2.0.1"
proto = {path = "../../lib/proto"}
poetry-grpc-plugin = "^0.1.1"
dataclasses-json = "^0.5.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
